+++ src/main/java/org/acme/graph/model/Vertex.java
@@ -26,18 +26,7 @@ public class Vertex {
 	 */
 	private Coordinate coordinate;
 
-	/**
-	 * dijkstra - coût pour atteindre le sommet
-	 */
-	private double cost;
-	/**
-	 * dijkstra - arc entrant avec le meilleur coût
-	 */
-	private Edge reachingEdge;
-	/**
-	 * dijkstra - indique si le sommet est visité
-	 */
-	private boolean visited;
+
 
 	/**
 	 * dijkstra - Arêtes entrantes
@@ -68,32 +57,6 @@ public class Vertex {
 
 	public void setCoordinate(Coordinate coordinate) {
 		this.coordinate = coordinate;
-	}
-
-	@JsonIgnore
-	public double getCost() {
-		return cost;
-	}
-
-	public void setCost(double cost) {
-		this.cost = cost;
-	}
-
-	@JsonIgnore
-	public Edge getReachingEdge() {
-		return reachingEdge;
-	}
-
-	public void setReachingEdge(Edge reachingEdge) {
-		this.reachingEdge = reachingEdge;
-	}
-
-	public boolean isVisited() {
-		return visited;
-	}
-
-	public void setVisited(boolean visited) {
-		this.visited = visited;
 	}
 	
 	public List<Edge> getInEdges() {
